* bebanjo: (61 commits)
  ui fixes:
  load more builds, display builds in table
  add spinner to job page
  update user_repos page
  update add project page
  redirect to "Add Project" page if user has no projects
  User can unsubscribe from project
  update vx-builder gem
  update rails to 4.0.8
  fixed \\r control code in terminal
  update vx-builder gem
  update vx-builder gem
  add last_month to company usage
  update vx-builder gem
  Specify the branch on the builds queue
  Use existing objects
  Bump vx-builder
  Add our changes again with the new design
  Remove deprecation warning is not needed to pass all message
  update vx-builder gem, fix typo in UsersCtrl
  ...

Conflicts:
	Gemfile
	Gemfile.lock
	app/models/build.rb
	app/models/project.rb
	db/schema.rb
	spec/controllers/api/status_controller_spec.rb
	spec/mailers/builds_mailer_spec.rb
	spec/models/build_notifier_spec.rb
	spec/models/company_usage_spec.rb
	spec/models/job_spec.rb
	spec/models/job_updater_spec.rb

+++ app/models/user.rb
@@ -17,13 +17,16 @@ class User < ActiveRecord::Base
     user_company.update(role: role)
   end
 
-  def default_company
-    companies.reorder("user_companies.default DESC").first
+  def default_company(reload = false)
+    if reload
+      @default_company = nil
+    end
+    @default_company ||= companies.reorder("user_companies.default DESC").first
   end
 
   def role(company)
     if company
-      user_companies.where(company_id: company.id).pluck(:role).first
+      user_companies.where(company: company).pluck(:role).first
     end
   end
 
@@ -41,9 +44,18 @@ class User < ActiveRecord::Base
         synced_repos = identity.sc.repos.map do |external_repo|
           UserRepo.find_or_create_by_sc company, identity, external_repo
         end
-        identity.user_repos.where("id NOT IN (?)", synced_repos.map(&:id)).each do |user_repo|
-          user_repo.destroy
+
+        collection =
+          if synced_repos.any?
+            identity.user_repos.where("id NOT IN (?)", synced_repos.map(&:id))
+          else
+            identity.user_repos.to_a
+          end
+
+        collection.each do |user_repo|
+          user_repo.destroy unless user_repo.project
         end
+
         synced_repos
       end
     end
@@ -55,6 +67,11 @@ class User < ActiveRecord::Base
 
   def add_to_company(company, role = 'developer')
     user_company = user_companies.find_or_initialize_by(company_id: company.id)
+
+    if user_company.persisted? and user_company.role == role
+      return true
+    end
+
     user_company.role = role
     if user_company.save
       user_company.default!
@@ -76,6 +93,15 @@ class User < ActiveRecord::Base
     end
   end
 
+  def update_with_company(company, params)
+    transaction do
+      if role = params.delete(:role)
+        add_to_company(company, role).or_rollback_transaction
+      end
+      update(params).or_rollback_transaction
+    end
+  end
+
   def set_default_company(company)
     user_companies.where(company_id: company.id).map(&:default!)
     touch
@@ -87,10 +113,11 @@ end
 #
 # Table name: users
 #
-#  id         :integer          not null, primary key
-#  email      :string(255)      not null
-#  name       :string(255)      not null
-#  created_at :datetime
-#  updated_at :datetime
+#  email       :string(255)      not null
+#  name        :string(255)      not null
+#  created_at  :datetime
+#  updated_at  :datetime
+#  back_office :boolean          default(FALSE)
+#  id          :uuid             not null, primary key
 #
